+++ NestSuite/NestSuite/NestSuiteShellWindow.FileSaveStateSync.cs
@@ -1,20 +1,4 @@
-using System.Collections.ObjectModel;
-using System.ComponentModel;
-using System.IO;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Input;
-using System.Windows.Documents;
-using System.Windows.Media;
 using NestSuite.ChatNest;
-using NestSuite.FileAssociation;
-using NestSuite.IdeaNest.ViewModels;
-using NestSuite.IdeaNest.Services;
-using NestSuite.NoteNest.Editor;
-using NestSuite.Services;
-using NestSuite.TempNest;
-using NestSuite.ViewModels;
-using NestSuite.Views;
 
 namespace NestSuite;